+++ openlyrics/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### New features
+
+- Added `LyricEntry::name` method.
+
 ## 0.1.0
 
 Initial release.